+++ src/QueryBin.d.ts
@@ -1,4 +1,4 @@
-export class QueryBin<T, Q extends QueryDefinitions<T> = QueryDefinitions<T>> {
+export class QueryBin<T, Q extends Queries<T> = QueryDefinitions<T>> {
   constructor(queries: Q, options?: Options);
 
   add(value: T): void;
@@ -26,32 +26,32 @@ export type Queries<T> = {
 };
 
 export type QueryDefinition<T> = {
-  queryAll: (all: T[]) => T[];
-  serializeForErrorMessage: (item: T) => string;
+  test: (item: T) => boolean;
+  serializeForErrorMessage?: (item: T) => string;
   noneFoundMessage: string;
   multipleFoundMessage: string;
 };
 
-export type QueryAll<T, Q extends QueryDefinitions<T>> = {
+export type QueryAll<T, Q extends Queries<T>> = {
   [Name in keyof Q]: (...args: Parameters<Q[Name]>) => T[];
 };
 
-export type Query<T, Q extends QueryDefinitions<T>> = {
+export type Query<T, Q extends Queries<T>> = {
   [Name in keyof Q]: (...args: Parameters<Q[Name]>) => T | null;
 };
 
-export type Get<T, Q extends QueryDefinitions<T>> = {
+export type Get<T, Q extends Queries<T>> = {
   [Name in keyof Q]: (...args: Parameters<Q[Name]>) => T;
 };
 
-export type GetAll<T, Q extends QueryDefinitions<T>> = {
+export type GetAll<T, Q extends Queries<T>> = {
   [Name in keyof Q]: (...args: Parameters<Q[Name]>) => T[];
 };
 
-export type FindAll<T, Q extends QueryDefinitions<T>> = {
+export type FindAll<T, Q extends Queries<T>> = {
   [Name in keyof Q]: (...args: Parameters<Q[Name]>) => Promise<T[]>;
 };
 
-export type Find<T, Q extends QueryDefinitions<T>> = {
+export type Find<T, Q extends Queries<T>> = {
   [Name in keyof Q]: (...args: Parameters<Q[Name]>) => Promise<T>;
 };