+++ src/QueryBin.js
@@ -5,16 +5,19 @@ export class QueryBin {
 
     this.latch = new Latch();
 
+    const queryAll = (queryDefinition) => {
+      return this.values.filter(queryDefinition.test);
+    };
+
     this.queryAll = mapValues(queries, (factory) => {
       return (...args) => {
-        const { queryAll } = factory(...args);
-        return queryAll(this.values);
+        return queryAll(factory(...args));
       };
     });
     this.query = mapValues(queries, (factory) => {
       return (...args) => {
         const queryDefinition = factory(...args);
-        const results = queryDefinition.queryAll(this.values);
+        const results = queryAll(queryDefinition);
         if (results.length === 0) return null;
         if (results.length > 1)
           throw this.multipleResultsError(queryDefinition, results);
@@ -24,7 +27,7 @@ export class QueryBin {
     this.get = mapValues(queries, (factory) => {
       return (...args) => {
         const queryDefinition = factory(...args);
-        const results = queryDefinition.queryAll(this.values);
+        const results = queryAll(queryDefinition);
         if (results.length === 0) throw this.noResultsError(queryDefinition);
         if (results.length > 1)
           throw this.multipleResultsError(queryDefinition, results);
@@ -34,7 +37,7 @@ export class QueryBin {
     this.getAll = mapValues(queries, (factory) => {
       return (...args) => {
         const queryDefinition = factory(...args);
-        const results = queryDefinition.queryAll(this.values);
+        const results = queryAll(queryDefinition);
         if (results.length === 0) throw this.noResultsError(queryDefinition);
         return results;
       };
@@ -43,7 +46,7 @@ export class QueryBin {
       return async (...args) => {
         const queryDefinition = factory(...args);
         return this.waitFor(() => {
-          const results = queryDefinition.queryAll(this.values);
+          const results = queryAll(queryDefinition);
           if (results.length === 0) throw this.noResultsError(queryDefinition);
           return results;
         });
@@ -53,7 +56,7 @@ export class QueryBin {
       return async (...args) => {
         const queryDefinition = factory(...args);
         return this.waitFor(() => {
-          const results = queryDefinition.queryAll(this.values);
+          const results = queryAll(queryDefinition);
           if (results.length === 0) throw this.noResultsError(queryDefinition);
           if (results.length > 1)
             throw this.multipleResultsError(queryDefinition, results);
@@ -137,17 +140,4 @@ class Latch {
       this.open = resolve;
     });
   }
-}
-
-function promiseWithResolvers() {
-  let resolve;
-  let reject;
-  const promise = new Promise((_resolve, _reject) => {
-    (resolve = _resolve), (reject = _reject);
-  });
-  return {
-    promise,
-    resolve,
-    reject,
-  };
 }