+++ src/example.test.ts
@@ -11,8 +11,7 @@ interface Request {
 
 function byMethodAndUrl(method: Method, url: string): QueryDefinition<Request> {
   return {
-    queryAll: (items) =>
-      items.filter((item) => item.method === method && item.url.includes(url)),
+    test: (item) => item.method === method && item.url.includes(url),
     serializeForErrorMessage: (item) => JSON.stringify(item, null, 2),
     noneFoundMessage: `Could not find requests with method ${method} and URL containing ${url}.`,
     multipleFoundMessage: `Multiple requests found method ${method} and URL containing ${url}.`,